templates分割 (#20)
